Cambios de Integer a Float en muestras

+++ src/main/java/ni/org/ics/zip/domain/Zp02BiospecimenCollection.java
@@ -26,29 +26,29 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     private String bscMatBldSpecify;
     private String bscMatBldTyp1;
     private String bscMatBldId1;
-    private Integer bscMatBldVol1;
+    private Float bscMatBldVol1;
     private String bscMatBldTyp2;
     private String bscMatBldId2;
-    private Integer bscMatBldVol2;
+    private Float bscMatBldVol2;
     private String bscMatBldTyp3;
     private String bscMatBldId3;
-    private Integer bscMatBldVol3;
+    private Float bscMatBldVol3;
     private String bscMatBldTyp4;
     private String bscMatBldId4;
-    private Integer bscMatBldVol4;
+    private Float bscMatBldVol4;
     private String bscMatBldTyp5;
     private String bscMatBldId5;
-    private Integer bscMatBldVol5;
+    private Float bscMatBldVol5;
     private String bscMatBldTyp6;
     private String bscMatBldId6;
-    private Integer bscMatBldVol6;
+    private Float bscMatBldVol6;
     private String bscMatBldTyp7;
     private String bscMatBldId7;
-    private Integer bscMatBldVol7;
+    private Float bscMatBldVol7;
     private String bscMatBldTyp8;
     private String bscMatBldId8;
-    private Integer bscMatBldVol8;
-    private Integer bscMatBldTotVol;
+    private Float bscMatBldVol8;
+    private Float bscMatBldTotVol;
     private String bscMatBldTime;
     private String bscMatBldCom;
     private String bscMatSlvaCol;
@@ -254,11 +254,11 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     }
 
     @Column(name = "bsc_mat_bld_vol_1", nullable = true)
-    public Integer getBscMatBldVol1() {
+    public Float getBscMatBldVol1() {
         return bscMatBldVol1;
     }
 
-    public void setBscMatBldVol1(Integer bscMatBldVol1) {
+    public void setBscMatBldVol1(Float bscMatBldVol1) {
         this.bscMatBldVol1 = bscMatBldVol1;
     }
 
@@ -281,11 +281,11 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     }
 
     @Column(name = "bsc_mat_bld_vol_2", nullable = true)
-    public Integer getBscMatBldVol2() {
+    public Float getBscMatBldVol2() {
         return bscMatBldVol2;
     }
 
-    public void setBscMatBldVol2(Integer bscMatBldVol2) {
+    public void setBscMatBldVol2(Float bscMatBldVol2) {
         this.bscMatBldVol2 = bscMatBldVol2;
     }
 
@@ -308,11 +308,11 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     }
 
     @Column(name = "bsc_mat_bld_vol_3", nullable = true)
-    public Integer getBscMatBldVol3() {
+    public Float getBscMatBldVol3() {
         return bscMatBldVol3;
     }
 
-    public void setBscMatBldVol3(Integer bscMatBldVol3) {
+    public void setBscMatBldVol3(Float bscMatBldVol3) {
         this.bscMatBldVol3 = bscMatBldVol3;
     }
 
@@ -335,11 +335,11 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     }
 
     @Column(name = "bsc_mat_bld_vol_4", nullable = true)
-    public Integer getBscMatBldVol4() {
+    public Float getBscMatBldVol4() {
         return bscMatBldVol4;
     }
 
-    public void setBscMatBldVol4(Integer bscMatBldVol4) {
+    public void setBscMatBldVol4(Float bscMatBldVol4) {
         this.bscMatBldVol4 = bscMatBldVol4;
     }
 
@@ -362,11 +362,11 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     }
 
     @Column(name = "bsc_mat_bld_vol_5", nullable = true)
-    public Integer getBscMatBldVol5() {
+    public Float getBscMatBldVol5() {
         return bscMatBldVol5;
     }
 
-    public void setBscMatBldVol5(Integer bscMatBldVol5) {
+    public void setBscMatBldVol5(Float bscMatBldVol5) {
         this.bscMatBldVol5 = bscMatBldVol5;
     }
 
@@ -389,11 +389,11 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     }
 
     @Column(name = "bsc_mat_bld_vol_6", nullable = true)
-    public Integer getBscMatBldVol6() {
+    public Float getBscMatBldVol6() {
         return bscMatBldVol6;
     }
 
-    public void setBscMatBldVol6(Integer bscMatBldVol6) {
+    public void setBscMatBldVol6(Float bscMatBldVol6) {
         this.bscMatBldVol6 = bscMatBldVol6;
     }
 
@@ -416,11 +416,11 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     }
 
     @Column(name = "bsc_mat_bld_vol_7", nullable = true)
-    public Integer getBscMatBldVol7() {
+    public Float getBscMatBldVol7() {
         return bscMatBldVol7;
     }
 
-    public void setBscMatBldVol7(Integer bscMatBldVol7) {
+    public void setBscMatBldVol7(Float bscMatBldVol7) {
         this.bscMatBldVol7 = bscMatBldVol7;
     }
 
@@ -443,20 +443,20 @@ public class Zp02BiospecimenCollection extends BaseMetaData implements Auditable
     }
 
     @Column(name = "bsc_mat_bld_vol_8", nullable = true)
-    public Integer getBscMatBldVol8() {
+    public Float getBscMatBldVol8() {
         return bscMatBldVol8;
     }
 
-    public void setBscMatBldVol8(Integer bscMatBldVol8) {
+    public void setBscMatBldVol8(Float bscMatBldVol8) {
         this.bscMatBldVol8 = bscMatBldVol8;
     }
 
     @Column(name = "bsc_mat_bld_tot_vol", nullable = true)
-    public Integer getBscMatBldTotVol() {
+    public Float getBscMatBldTotVol() {
         return bscMatBldTotVol;
     }
 
-    public void setBscMatBldTotVol(Integer bscMatBldTotVol) {
+    public void setBscMatBldTotVol(Float bscMatBldTotVol) {
         this.bscMatBldTotVol = bscMatBldTotVol;
     }
